Fix the seed files: row_number returns bigint, not int
Reported from Supabase:

  ERROR: 42883: operator does not exist: date - bigint
  QUERY: insert into fleet_vehicle ...

row_number() returns bigint. PostgreSQL will not subtract a bigint from a
date, so the vehicle purchase_date expression failed and took the whole seed
with it.

It would have failed twice more after that — the maintenance scheduled_date
and completed_date use the same value — and once more in a way that is easy
to miss: an array subscript also requires an integer, so picking a site name
with [1 + (fv.ord % 2)] would have failed too.

Cast at the source rather than at each use: all eight row_number() calls now
return int, which fixes every downstream use at once.

Audited the rest rather than assume this was the only one. Every date
expression and every array subscript in the generated SQL now has integer
operands: i, d and k come from generate_series, v_days is declared int, and
the row numbers are cast. Integer columns — odometer, orders_processed,
quantity, mileage, quantity_reserved — all receive integers.

The seed opens by deleting its own company, so a failed run leaves nothing
behind and re-running is clean.

+++ supabase/seed/SEED_automotive_parts.sql
@@ -85,7 +85,7 @@ begin
     round((61 + ((d * 11 + ws.ord * 19) % 34))::numeric, 2)
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from warehouse_site where company_id = v_company
   ) as ws;
 
@@ -104,7 +104,7 @@ begin
     current_date - (400 + (v.rn * 53) % 1800),
     0
   from (
-    select vtype, fuel, n, row_number() over () as rn
+    select vtype, fuel, n, row_number() over ()::int as rn
     from (values
       ('Forklift (counterbalance)', 'Battery electric', 1),
       ('Forklift (counterbalance)', 'Battery electric', 2),
@@ -150,7 +150,7 @@ begin
     case when (d + fv.ord) % 19 = 0 then 'cancelled' else 'completed' end
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from fleet_vehicle
     where company_id = v_company and status = 'active'
   ) as fv
@@ -177,7 +177,7 @@ begin
       'Depot Workshop ' || (1 + (fv.ord % 3)),
       'WO-' || lpad(((fv.ord * 10) + k)::text, 5, '0')
     from (
-      select id, row_number() over (order by name) as ord
+      select id, row_number() over (order by name)::int as ord
       from fleet_vehicle where company_id = v_company
     ) as fv
     cross join generate_series(1, 3) as k
@@ -233,7 +233,7 @@ begin
     current_date - d
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by sku) as ord
+    select id, row_number() over (order by sku)::int as ord
     from inventory_sku where company_id = v_company limit 300
   ) as s
   where (d + s.ord) % 7 = 0;
@@ -265,7 +265,7 @@ begin
     'TXN-' || lpad(((d * 20) + cc.ord)::text, 6, '0')
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by code) as ord
+    select id, row_number() over (order by code)::int as ord
     from finance_cost_center where company_id = v_company
   ) as cc
   where (d + cc.ord) % 2 = 0;
@@ -284,7 +284,7 @@ begin
     round((6.5 + ((d + e.ord) % 4))::numeric, 2)
   from generate_series(0, 29) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from hr_employee where company_id = v_company and status = 'active'
   ) as e;
 
@@ -300,7 +300,7 @@ begin
     'Recorded as part of the quarterly review cycle.',
     'Line Manager'
   from (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from hr_employee where company_id = v_company and status = 'active'
   ) as e
   where e.ord % 4 <> 0;
