+++ supabase/seed/SEED_third_party_logistics.sql
@@ -86,7 +86,7 @@ begin
     round((61 + ((d * 11 + ws.ord * 19) % 34))::numeric, 2)
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from warehouse_site where company_id = v_company
   ) as ws;
 
@@ -105,7 +105,7 @@ begin
     current_date - (400 + (v.rn * 53) % 1800),
     0
   from (
-    select vtype, fuel, n, row_number() over () as rn
+    select vtype, fuel, n, row_number() over ()::int as rn
     from (values
       ('Tractor unit', 'Diesel', 1),
       ('Tractor unit', 'Diesel', 2),
@@ -167,7 +167,7 @@ begin
     case when (d + fv.ord) % 19 = 0 then 'cancelled' else 'completed' end
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from fleet_vehicle
     where company_id = v_company and status = 'active'
   ) as fv
@@ -194,7 +194,7 @@ begin
       'Depot Workshop ' || (1 + (fv.ord % 3)),
       'WO-' || lpad(((fv.ord * 10) + k)::text, 5, '0')
     from (
-      select id, row_number() over (order by name) as ord
+      select id, row_number() over (order by name)::int as ord
       from fleet_vehicle where company_id = v_company
     ) as fv
     cross join generate_series(1, 3) as k
@@ -248,7 +248,7 @@ begin
     current_date - d
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by sku) as ord
+    select id, row_number() over (order by sku)::int as ord
     from inventory_sku where company_id = v_company limit 300
   ) as s
   where (d + s.ord) % 7 = 0;
@@ -280,7 +280,7 @@ begin
     'TXN-' || lpad(((d * 20) + cc.ord)::text, 6, '0')
   from generate_series(0, v_days - 1) as d
   cross join (
-    select id, row_number() over (order by code) as ord
+    select id, row_number() over (order by code)::int as ord
     from finance_cost_center where company_id = v_company
   ) as cc
   where (d + cc.ord) % 2 = 0;
@@ -299,7 +299,7 @@ begin
     round((6.5 + ((d + e.ord) % 4))::numeric, 2)
   from generate_series(0, 29) as d
   cross join (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from hr_employee where company_id = v_company and status = 'active'
   ) as e;
 
@@ -315,7 +315,7 @@ begin
     'Recorded as part of the quarterly review cycle.',
     'Line Manager'
   from (
-    select id, row_number() over (order by name) as ord
+    select id, row_number() over (order by name)::int as ord
     from hr_employee where company_id = v_company and status = 'active'
   ) as e
   where e.ord % 4 <> 0;
